requirements.txt
fastapi>=0.78
uvicorn[standard]>=0.18
streamlit>=1.20
requests>=2.0
beautifulsoup4>=4.6
transformers>=4.25
torch>=2.0
kafka-python>=2.0
openai>=0.27
pandas>=1.3
flake8>=4.0
pytest>=7.0
